+++ src/lib/htmlBenchmarkFormatter.ts
@@ -2,56 +2,16 @@ import fs from "fs";
 import { BenchmarkResultsJSONFormat } from "./benchmarkManager";
 import { HTMLFormatter } from "./htmlFormatter";
 import { JSONFormatter } from "./jsonFormatter";
-import { RangesTuple } from "./summary";
 
-export class HTMLBenchmarkFormatter extends HTMLFormatter {
-  public static makeBody(jsonString: string): string {
-    const jsonData: BenchmarkResultsJSONFormat[] = JSON.parse(
-      jsonString,
-      JSONFormatter.JSONReviverFunction,
-    );
-    const tableRows: string[] = new Array();
-    const comparisonPages: string[] = new Array();
+export class HTMLBenchmarkFormatter extends HTMLFormatter<BenchmarkResultsJSONFormat> {
+  public toComparePage(
+    matchedFile: string,
+    matchingFile: string,
+    benchmarkResults: BenchmarkResultsJSONFormat,
+  ): string {
+    const comparePage: string = this.toCompareView(matchedFile, matchingFile, benchmarkResults);
 
-    for (const benchmarkResults of jsonData.values()) {
-      tableRows.push(
-        this.makeTableRow(
-          benchmarkResults.matchedFile,
-          benchmarkResults.matchingFile,
-          benchmarkResults.benchmarkResults.falseRangesTuples.concat(
-            benchmarkResults.benchmarkResults.matchingRangesTuples,
-          ),
-        ),
-      );
-      comparisonPages.push(this.toViewPage(benchmarkResults));
-    }
-    return (
-      `<div>` +
-      `<p>Dolos summary</p>` +
-      `<p>${new Date().toUTCString()}` +
-      `<hr>` +
-      `</div>` +
-      `<div id="Index">\n` +
-      `<table>\n` +
-      `<tbody>\n` +
-      `<tr>\n` +
-      `<th class="filename-column">File 1</th>\n` +
-      `<th class="filename-column">File 2</th>\n` +
-      `<th class="lines-matched-column">Lines matched in File 1</th>\n` +
-      `<th class="lines-matched-column">Lines matched in File 2</th>\n` +
-      `</tr>\n` +
-      `${tableRows.join("\n")}\n` +
-      `</tbody>\n` +
-      `</table>\n` +
-      `</div>\n` +
-      `${comparisonPages.join("\n")}`
-    );
-  }
-
-  private static toViewPage(benchmarkResults: BenchmarkResultsJSONFormat): string {
-    const comparePage: string = this.toBenchmarkView(benchmarkResults);
-
-    const id: string = this.makeId(benchmarkResults.matchedFile, benchmarkResults.matchingFile);
+    const id: string = this.makeId(matchedFile, matchingFile);
     return (
       `<div style="display:none" id="${id}">\n` +
       `<div> <a href=# onclick="return swap('Index', '${id}')">Back to index</a> </div>\n` +
@@ -59,56 +19,25 @@ export class HTMLBenchmarkFormatter extends HTMLFormatter {
       `</div>\n`
     );
   }
-  private static sortRangesTuples([r11, r12]: RangesTuple, [r21, r22]: RangesTuple): number {
-    const diff = r11.from - r21.from;
-    if (diff === 0) {
-      return r12.from - r22.from;
-    } else {
-      return diff;
-    }
-  }
 
-  private static toMarkingDivAndToggleButton(
+  public toCompareView(
+    matchedFile: string,
+    matchingFile: string,
     benchmarkResults: BenchmarkResultsJSONFormat,
-    rightMarkedAreas: string[],
-    leftMarkedAreas: string[],
-    rangesTuples: RangesTuple[],
-    rangesList: string[],
-    colour: string,
-    indexOffset: number = 0,
-  ) {
-    for (const [index, [leftRange, rightRange]] of rangesTuples.entries()) {
-      const id: string = `${HTMLFormatter.makeId(
-        benchmarkResults.matchedFile,
-        benchmarkResults.matchingFile,
-        index + indexOffset,
-      )}`;
-      leftMarkedAreas.push(
-        HTMLFormatter.rangeToMarkingDiv(leftRange, `background: ${colour}`, id),
-      );
-      rightMarkedAreas.push(
-        HTMLFormatter.rangeToMarkingDiv(rightRange, `background: ${colour}`, id),
-      );
-      const rangesTupleString: string = `[${leftRange.toString()}, ${rightRange.toString()}]`;
-      rangesList.push(
-        `<div class="range" style="color: ${colour};" >` +
-          `<input type="checkbox" class="checkbox" data="${id}">` +
-          `${this.escapeHtml(rangesTupleString)}` +
-          `</div>`,
-      );
-    }
-  }
-
-  private static toBenchmarkView(benchmarkResults: BenchmarkResultsJSONFormat): string {
-    const description: string = this.escapeHtml(
-      `${benchmarkResults.matchedFile} => ${benchmarkResults.matchingFile}`,
+  ): string {
+    const description: string = HTMLBenchmarkFormatter.escapeHtml(
+      `${matchedFile} => ${matchingFile}`,
     );
-    benchmarkResults.benchmarkResults.falseRangesTuples.sort(this.sortRangesTuples);
-    benchmarkResults.benchmarkResults.matchingRangesTuples.sort(this.sortRangesTuples);
-    benchmarkResults.expected.sort(this.sortRangesTuples);
+    benchmarkResults.benchmarkResults.falseRangesTuples.sort(
+      HTMLBenchmarkFormatter.sortRangesTuples,
+    );
+    benchmarkResults.benchmarkResults.matchingRangesTuples.sort(
+      HTMLBenchmarkFormatter.sortRangesTuples,
+    );
+    benchmarkResults.expected.sort(HTMLBenchmarkFormatter.sortRangesTuples);
 
-    const right: string = fs.readFileSync(benchmarkResults.matchingFile, "utf8");
-    const left: string = fs.readFileSync(benchmarkResults.matchedFile, "utf8");
+    const right: string = fs.readFileSync(matchingFile, "utf8");
+    const left: string = fs.readFileSync(matchedFile, "utf8");
 
     const leftMarkedAreas: string[] = [];
     const rightMarkedAreas: string[] = [];
@@ -117,8 +46,9 @@ export class HTMLBenchmarkFormatter extends HTMLFormatter {
     const matchingRangesTuples: string[] = [];
 
     let indexOffset: number = 0;
-    this.toMarkingDivAndToggleButton(
-      benchmarkResults,
+    HTMLBenchmarkFormatter.toMarkingDivAndToggleButton(
+      benchmarkResults.matchedFile,
+      benchmarkResults.matchingFile,
       rightMarkedAreas,
       leftMarkedAreas,
       benchmarkResults.benchmarkResults.falseRangesTuples,
@@ -129,7 +59,8 @@ export class HTMLBenchmarkFormatter extends HTMLFormatter {
     indexOffset += benchmarkResults.benchmarkResults.falseRangesTuples.length;
 
     this.toMarkingDivAndToggleButton(
-      benchmarkResults,
+      benchmarkResults.matchedFile,
+      benchmarkResults.matchingFile,
       rightMarkedAreas,
       leftMarkedAreas,
       benchmarkResults.benchmarkResults.matchingRangesTuples,
@@ -141,7 +72,8 @@ export class HTMLBenchmarkFormatter extends HTMLFormatter {
     indexOffset += benchmarkResults.benchmarkResults.matchingRangesTuples.length;
 
     this.toMarkingDivAndToggleButton(
-      benchmarkResults,
+      benchmarkResults.matchedFile,
+      benchmarkResults.matchingFile,
       rightMarkedAreas,
       leftMarkedAreas,
       benchmarkResults.expected,
@@ -183,6 +115,54 @@ export class HTMLBenchmarkFormatter extends HTMLFormatter {
       `</pre>\n` +
       `</div>\n` +
       `</div>\n`
+    );
+  }
+  public makeBody(jsonString: string): string {
+    const jsonData: BenchmarkResultsJSONFormat[] = JSON.parse(
+      jsonString,
+      JSONFormatter.JSONReviverFunction,
+    );
+    const tableRows: string[] = new Array();
+    const comparisonPages: string[] = new Array();
+
+    for (const benchmarkResults of jsonData.values()) {
+      tableRows.push(
+        HTMLBenchmarkFormatter.makeTableRow(
+          benchmarkResults.matchedFile,
+          benchmarkResults.matchingFile,
+          benchmarkResults.benchmarkResults.falseRangesTuples.concat(
+            benchmarkResults.benchmarkResults.matchingRangesTuples,
+          ),
+        ),
+      );
+      comparisonPages.push(
+        this.toComparePage(
+          benchmarkResults.matchedFile,
+          benchmarkResults.matchingFile,
+          benchmarkResults,
+        ),
+      );
+    }
+    return (
+      `<div>` +
+      `<p>Dolos summary</p>` +
+      `<p>${new Date().toUTCString()}` +
+      `<hr>` +
+      `</div>` +
+      `<div id="Index">\n` +
+      `<table>\n` +
+      `<tbody>\n` +
+      `<tr>\n` +
+      `<th class="filename-column">File 1</th>\n` +
+      `<th class="filename-column">File 2</th>\n` +
+      `<th class="lines-matched-column">Lines matched in File 1</th>\n` +
+      `<th class="lines-matched-column">Lines matched in File 2</th>\n` +
+      `</tr>\n` +
+      `${tableRows.join("\n")}\n` +
+      `</tbody>\n` +
+      `</table>\n` +
+      `</div>\n` +
+      `${comparisonPages.join("\n")}`
     );
   }
 }
